# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText
  description: |
    API for WASAText, which is a online messaging application
    similar to whatsapp
  version: 1.0.0
...
security:
  - bearerAuth: [ ]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    message:
      description: a single message
      type: object
      properties:
        id:
          type: integer
          description: identifier for message
        status:
          description: status of the message
          type: string
          enum: [ unknown, received, read ]
        messageContent:
          description: the content of the message, string or a file
          type: string
        comment:
          description: emoticon for the message
          type: array
          items:
            type: integer
        timestamp:
          description: time when the message was sent
          type: string
          format: date-time
          example: '2019-10-19 17:12:30'
        sender:
          description: user who sent the message
          type: integer
        conversation:
          description: conversation that the message belongs to
          type: integer
...
    user:
      description: a single user
      type: object
      properties:
        id:
          type: integer
          description: identifier for user
        name:
          description: name of the user
          type: string
        picture:
          description: picture of the user
          type: string
        conversations:
          description: list of all conversations
          type: array
          items:
            type: integer
    comment:
      description: a comment for the message
      type: object
      properties:
        id:
          type: integer
          description: identifier for comment
        commentContent:
          description: emoticon
        user:
          description: user who added the comment
          type: integer
  parameters:
    userId:
      description: identifier for the user
      name: Id
      in: path
      required: true
      schema:
        type: integer
    conversationId:
      description: identifier for the conversation
      name: conversationId
      in: path
      required: true
      schema:
        type: integer
    messageId:
      description: identifier for the message
      name: messageId
      in: path
      required: true
      schema:
        type: integer
    commentId:
      description: identifier for the comment
      name: commentId
      in: path
      required: true
      schema:
        type: integer
paths:
  /session:
    post:
      security: []
      tags:
        - login
      summary: Logs in the user
      description: >-
        if the user does not exist, it will be created and an identifier is
        returned.

        IF the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Maria
                  pattern: ^.*?$
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: string
                    example: abcdef012345
  /mainpage/{Id}:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      tags:
        - users
      summary: gets user
      operationId: getSelf
      description: get the user
      responses:
        "200":
          description: user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
  /mainpage/{Id}/users:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      tags:
        - users
      summary: gets users
      operationId: getUsers
      description: lists all the users
      responses:
        "200":
          description: an array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /mainpage/{Id}/settings/name:
    parameters:
      - $ref: "#/components/parameters/userId"
    put:
      tags:
        - users
      summary: changes name
      operationId: setMyUserName
      description: change the name of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string

      responses:
        "204":
          description: successfully changed the name
        "400":
          description: could not change into this name
  /mainpage/{Id}/settings/picture:
    parameters:
      - $ref: "#/components/parameters/userId"
    put:
      tags:
        - users
      summary: changes picture
      operationId: setMyPhoto
      description: change the picture of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "204":
          description: successfully changed the picture
        "400":
          description: could not change the picture
...
  /mainpage/{Id}/conversations/{conversationId}//{messageId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    delete:
      tags:
        - messages
      summary: deletes message
      operationId: deleteMessage
      description: delete my message
      responses:
        "200":
          description: message successfully deleted
          content:
            application/json:
              schema:
                type: string
        "404":
          description: message not found
    get:
      tags:
        - messages
      summary: gets message
      operationId: getMessage
      description: get a specific message
      responses:
        "200":
          description: object message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/message"
        "404":
          description: message not found
          content:
            application/json:
              schema:
                type: integer
                enum: [ 0 ]
    post:
      tags:
        - messages
      summary: forwards message
      operationId: forwardMessage
      description: forward the message to other conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "201":
          description: message successfully forwarded
          content:
            application/json:
              schema:
                type: integer
        "404":
          description: could not find the user
  /mainpage/{Id}/conversations/{conversationId}//{messageId}/comments:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    get:
      tags:
        - comments
      summary: gets comments
      operationId: getComments
      description: get all comments for this message
      responses:
        "200":
          description: array of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/comment"
    post:
      tags:
        - comments
      summary: makes comments
      operationId: commentMessage
      description: add a new comment on this message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        "201":
          description: comment successfully added
          content:
            application/json:
              schema:
                type: integer
  /mainpage/{Id}/conversations/{conversationId}//{messageId}/comments/{commentId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
      - $ref: "#/components/parameters/commentId"
    get:
      tags:
        - comments
      summary: gets comment
      operationId: getSpecificComment
      description: get a specific comment for this message
      responses:
        "200":
          description: a comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/comment"
        "404":
          description: comment not found
          content:
            application/json:
              schema:
                type: integer
                enum: [ 0 ]
    delete:
      tags:
        - comments
      summary: deletes comment
      operationId: uncommentMessage
      description: delete my comment
      responses:
        "200":
          description: comment successfully deleted
          content:
            application/json:
              schema:
                type: string
        "404":
          description: comment not found
...